# SYCL: Add set_rows support for quantized types 