--- a/ggml/src/ggml-sycl/ggml-sycl.cpp
+++ b/ggml/src/ggml-sycl/ggml-sycl.cpp
@@ -4385,10 +4385,17 @@ static bool ggml_backend_sycl_device_supports_op(ggml_backend_dev_t dev, const g
             }
         case GGML_OP_SET_ROWS:
             {
-                // TODO: add support
-                // ref: https://github.com/ggml-org/llama.cpp/pull/14274
-#pragma message("TODO: implement BF16, Q4_0, Q4_1, Q5_0, Q5_1, Q8_0, IQ4_NL support (https://github.com/ggml-org/llama.cpp/pull/14661)")
-                return (op->type == GGML_TYPE_F32 || (op->type == GGML_TYPE_F16 && op->src[0]->type == GGML_TYPE_F32 && op->src[1]->type == GGML_TYPE_I64));
+                return ((
+                            op->type == GGML_TYPE_F32 ||
+                            op->type == GGML_TYPE_F16 ||
+                            op->type == GGML_TYPE_BF16 ||
+                            op->type == GGML_TYPE_Q8_0 ||
+                            op->type == GGML_TYPE_Q5_1 ||
+                            op->type == GGML_TYPE_Q5_0 ||
+                            op->type == GGML_TYPE_Q4_1 ||
+                            op->type == GGML_TYPE_Q4_0 ||
+                            op->type == GGML_TYPE_IQ4_NL
+                            ) && (op->src[1]->type == GGML_TYPE_I64));
             } break;
         case GGML_OP_CPY:
             {