--- a/ggml/src/ggml-sycl/ggml-sycl.cpp
+++ b/ggml/src/ggml-sycl/ggml-sycl.cpp
@@ -4385,18 +4385,12 @@ static bool ggml_backend_sycl_device_supports_op(ggml_backend_dev_t dev, const g
             }
         case GGML_OP_SET_ROWS:
             {
-                return ((
-                            op->type == GGML_TYPE_F32 ||
-                            op->type == GGML_TYPE_F16 ||
-                            op->type == GGML_TYPE_BF16 ||
-                            op->type == GGML_TYPE_Q8_0 ||
-                            op->type == GGML_TYPE_Q5_1 ||
-                            op->type == GGML_TYPE_Q5_0 ||
-                            op->type == GGML_TYPE_Q4_1 ||
-                            op->type == GGML_TYPE_Q4_0 ||
-                            op->type == GGML_TYPE_IQ4_NL
-                            ) && (op->src[1]->type == GGML_TYPE_I64));
-            } break;
+                return ((op->type == GGML_TYPE_F32 || op->type == GGML_TYPE_F16 || op->type == GGML_TYPE_BF16 ||
+                         op->type == GGML_TYPE_Q8_0 || op->type == GGML_TYPE_Q5_1 || op->type == GGML_TYPE_Q5_0 ||
+                         op->type == GGML_TYPE_Q4_1 || op->type == GGML_TYPE_Q4_0 || op->type == GGML_TYPE_IQ4_NL) &&
+                        (op->src[1]->type == GGML_TYPE_I64));
+            }
+            break;
         case GGML_OP_CPY:
             {
                 ggml_type src0_type = op->src[0]->type;