--- a/ggml/src/ggml-sycl/set_rows.cpp
+++ b/ggml/src/ggml-sycl/set_rows.cpp
@@ -29,7 +29,7 @@ static void set_rows_sycl_q(const char * __restrict__ src0_d,
                             const int64_t ne02,
                             const int64_t ne03,
                             const int64_t ne10,
-                            int64_t       ne11,
+                            const int64_t ne11,
                             const int64_t ne12,
                             const int64_t ne13,
                             // strides for src0
@@ -66,11 +66,11 @@ static void set_rows_sycl_q(const char * __restrict__ src0_d,
         const int64_t i12         = i03 % ne12;
         const int64_t i11         = i02 % ne11;
         const int64_t i10         = i01;
-        size_t        src_offset  = calculate_offset<3>({ nb01, nb02, nb03 }, { i01, i02, i03 });
+        const size_t  src_offset  = calculate_offset<3>({ nb01, nb02, nb03 }, { i01, i02, i03 });
         const char *  src_block   = src0_d + src_offset + i00 * sizeof(float);
-        size_t        src1_offset = calculate_offset<3>({ nb10, nb11, nb12 }, { i10, i11, i12 });
+        const size_t  src1_offset = calculate_offset<3>({ nb10, nb11, nb12 }, { i10, i11, i12 });
         const int64_t dst_row     = src1_d[src1_offset / sizeof(int64_t)];
-        size_t        dst_offset =
+        const size_t  dst_offset =
             calculate_offset<3>({ nb1, nb2, nb3 }, { dst_row, i02, i03 }) + (i00 / qk) * sizeof(blockType);
         char * dst_block = reinterpret_cast<char *>(reinterpret_cast<char *>(dst_d) + dst_offset);
         cpyblck(src_block, dst_block);